[FIX] account_banking_ach_discount: decide grouping criteria
